requirements.txt
# Python deps
psycopg2-binary==2.9.9
sqlalchemy==2.0.30
python-dotenv==1.0.1
faker==25.0.0

# AI
anthropic==0.28.0

# API
fastapi==0.111.0
uvicorn[standard]==0.30.0
pydantic==2.7.1
apscheduler==3.10.4
httpx==0.27.0

# CLI (optional Typer for scripts)
typer==0.12.3

# Testing
pytest==8.2.0
pytest-asyncio==0.23.6
